Add generic `IgnoreEcallSystemInstructionHandler`
